Split up code more nicely
