+++ src/bin/commands/login.ts
@@ -1,27 +1,11 @@
-import toml from '@iarna/toml';
-import fs from 'fs/promises';
 import getPort from 'get-port';
 import http from 'http';
-import ini from 'ini';
 import open from 'open';
-import os from 'os';
-import path from 'path';
 
-import { CACHE_DIR, CACHE_DIR_FILE } from '../utils';
+import { storeSession, storeSessionForBun, storeSessionforNPM } from '../utils/session';
 
 export default async () => {
   console.log('Logging in...');
-
-  // Ensure that the cache directory exists.
-  try {
-    await fs.stat(CACHE_DIR);
-  } catch (err) {
-    if ((err as { code: string }).code === 'ENOENT') {
-      await fs.mkdir(CACHE_DIR);
-    } else {
-      throw err;
-    }
-  }
 
   let token: string | undefined;
 
@@ -61,69 +45,8 @@ export default async () => {
     server.close();
   }
 
-  await fs.writeFile(
-    CACHE_DIR_FILE,
-    JSON.stringify(
-      {
-        token,
-      },
-      null,
-      2,
-    ),
-  );
-
-  const authenticateNPM = async () => {
-    const npmConfigFile = process.env['npm_config_userconfig'] || path.join(os.homedir(), '.npmrc');
-
-    let npmConfigContents: Record<string, any> = {};
-
-    try {
-      const contents = await fs.readFile(npmConfigFile, { encoding: 'utf-8' });
-      npmConfigContents = ini.parse(contents);
-    } catch (err) {
-      if ((err as { code: string }).code !== 'ENOENT') {
-        console.error(`Failed to read npm config file at ${npmConfigFile}`, err);
-      }
-    }
-
-    npmConfigContents['@ronin:registry'] = 'https://ronin.supply';
-    npmConfigContents['//ronin.supply/:_authToken'] = token;
-
-    await fs.writeFile(npmConfigFile, ini.stringify(npmConfigContents), {
-      encoding: 'utf-8',
-    });
-  };
-
-  authenticateNPM();
-
-  const authenticateBun = async () => {
-    const bunConfigFile = path.join(os.homedir(), '.bunfig.toml');
-
-    let bunConfigContents: Record<string, any> = {};
-
-    try {
-      const contents = await fs.readFile(bunConfigFile, { encoding: 'utf-8' });
-      bunConfigContents = toml.parse(contents);
-    } catch (err) {
-      if ((err as { code: string }).code !== 'ENOENT') {
-        console.error(`Failed to read npm config file at ${bunConfigFile}`, err);
-      }
-    }
-
-    // Safely initialize potentially missing keys.
-    if (!bunConfigContents.install) bunConfigContents.install = {};
-    if (!bunConfigContents.install.scopes) bunConfigContents.install.scopes = {};
-    if (!bunConfigContents.install.scopes.ronin) bunConfigContents.install.scopes.ronin = {};
-
-    bunConfigContents.install.scopes.ronin.url = 'https://ronin.supply';
-    bunConfigContents.install.scopes.ronin.token = token;
-
-    await fs.writeFile(bunConfigFile, toml.stringify(bunConfigContents), {
-      encoding: 'utf-8',
-    });
-  };
-
-  authenticateBun();
+  // Initialize sessions for the RONIN CLI and all relevant package managers.
+  await Promise.all([storeSession(token), storeSessionforNPM(token), storeSessionForBun(token)]);
 
   console.log('Successfully logged in');
 };
